feature: extract formio initialization code
