+++ src/main/app/src/app/taken/model/taak.ts
@@ -9,6 +9,7 @@ import { TaakRechten } from "../../policy/model/taak-rechten";
 import { GeneratedType } from "../../shared/utils/generated-types";
 import { TaakStatus } from "./taak-status.enum";
 import { Taakinformatie } from "./taakinformatie";
+import {FormioForm} from "@formio/angular";
 
 export class Taak {
   id: string;
@@ -25,7 +26,7 @@ export class Taak {
   status: TaakStatus;
   formulierDefinitieId: FormulierDefinitieID;
   formulierDefinitie: FormulierDefinitie;
-  formioFormulier: unknown;
+  formioFormulier: FormioForm;
   tabellen: Record<string, string[]>;
   taakdata: Record<string, string>;
   taakinformatie: Taakinformatie;
